+++ backend/src/main/resources/db/migration/V39__create_async_task_failure_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE async_task_failure
     target_type   VARCHAR(20)  NOT NULL,
     target_id     VARCHAR(100) NOT NULL,
     error_message TEXT         NOT NULL,
-    is_retryable  BOOLEAN      NOT NULL,
+    is_retryable  BIT(1)      NOT NULL,
     retry_count   INT          NOT NULL DEFAULT 0,
     status        VARCHAR(20)  NOT NULL,
     created_at    DATETIME(6)  NOT NULL,
